requirements.txt
openai
google-generativeai
python-dotenv
flask
gunicorn
python-dotenv